issue #131 refactor: common design component directory change
