+++ src/components/common/Design/Box/Box.styles.ts
