+++ src/components/common/Design/Carousel/Carousel.style.ts
