+++ src/components/common/Design/Carousel/Carousel.tsx
@@ -5,9 +5,9 @@ import LeftArrowIcon from "@/assets/svg/ic-left-arrow.svg?react";
 import RightArrowIcon from "@/assets/svg/ic-right-arrow.svg?react";
 
 import { Box } from "@/components/common";
-import CarouselItem from "@/components/common/Carousel/CarouselItem";
-import Dots from "@/components/common/Carousel/Dots";
-import Gallery from "@/components/common/Carousel/Gallery/Gallery";
+import CarouselItem from "@/components/common/Design/Carousel/CarouselItem";
+import Dots from "@/components/common/Design/Carousel/Dots";
+import Gallery from "@/components/common/Design/Carousel/Gallery/Gallery";
 
 import useCarousel from "@/hooks/useCarousel";
 import useClickOutSide from "@/hooks/useClickOutSide";
