+++ src/components/common/Design/Carousel/Dots.tsx
