+++ src/components/common/Design/Carousel/Gallery/Gallery.style.ts
