+++ src/components/common/Design/Carousel/Gallery/GallerySlider/GallerySlider.style.ts
