+++ src/components/common/Design/Dropdown/DropdownList.tsx
@@ -1,11 +1,11 @@
 import { useContext } from "react";
 
-import Box from "@/components/common/Box/Box";
-import { DropdownProvider } from "@/components/common/Dropdown/Dropdown";
+import Box from "@/components/common/Design/Box/Box";
+import { DropdownProvider } from "@/components/common/Design/Dropdown/Dropdown";
 
 import type { DropdonwListType } from "@/types/common";
 
-import { dropdownListStyle } from "@/components/common/Dropdown/Dropdown.style";
+import { dropdownListStyle } from "@/components/common/Design/Dropdown/Dropdown.style";
 
 const DropdownList = ({ children, ...props }: DropdonwListType) => {
 	const { isDropdownOpen } = useContext(DropdownProvider);
