+++ src/components/common/Design/Skeleton/Skeleton.style.ts
