+++ src/components/common/Design/Text/Text.tsx
@@ -2,7 +2,7 @@ import type { ComponentPropsWithoutRef } from "react";
 
 import type { Size } from "@/types/size";
 
-import { getSizeStyling } from "@/components/common/Text/Text.style";
+import { getSizeStyling } from "@/components/common/Design/Text/Text.style";
 
 export interface TextProps extends ComponentPropsWithoutRef<"p"> {
 	size?: Extract<Size, "xSmall" | "small" | "medium" | "large" | "xLarge">;
