+++ src/stories/Box.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Box from "@/components/common/Box/Box";
+import Box from "@/components/common/Design/Box/Box";
 import { containerStyle } from "@/stories/styles";
 
 import { Theme } from "@/styles/Theme";
